Add files via upload
The web application represents the prototype for viewing, adding, updating and deleting customers
details of an organization. The application is developed using the frontend language Angualr.JS and
the backend language C# on .Net Framework with Microsoft SQL Server 2016 Express LocalDB as the
database. There are mainly 3 pages in the application- Home page, Customer Data which displays all
the registered customers data and Create Customer where new customers data is entered. The
application also allows the user Edit and Delete a specific data as well.